+++ src/account/account.module.ts
@@ -1,8 +1,8 @@
-import { AccountRestController } from './controllers-rest/account.rest.controller';
-import { CreateAccountRestService } from './controllers-rest/services-create/create.account.rest.service';
-import { DepositAccountRestService } from './controllers-rest/services-deposit/deposit.account.rest.service';
-import { WithdrawAccountRestService } from './controllers-rest/services-withdraw/withdraw.account.rest.service';
-import { MapAccountRepositoryService } from './infra-repository/map.account.repository.service';
+import { AccountRestController } from './controllers/rest/account.rest.controller';
+import { CreateAccountRestService } from './controllers/rest/services/create/create.account.rest.service';
+import { DepositAccountRestService } from './controllers/rest/services/deposit/deposit.account.rest.service';
+import { WithdrawAccountRestService } from './controllers/rest/services/withdraw/withdraw.account.rest.service';
+import { MapAccountRepositoryService } from './infra/repository/map.account.repository.service';
 import { AccountService } from './services/account/account.service';
 import { CreateAccountService } from './services/create/create.account.service';
 import { DepositAccountService } from './services/deposit/deposit.account.service';
