+++ src/account/controllers/rest/services/create/create.account.rest.service.dtos.ts
@@ -1,4 +1,4 @@
-import { AccountData } from '../../domain/entity/account.entity.data';
+import { AccountData } from '../../../../domain/entity/account.entity.data';
 
 export type CreateAccountRestServiceParamsDTO = {
   nickname: string;
