+++ src/account/controllers/rest/services/withdraw/withdraw.account.rest.service.dtos.ts
@@ -1,4 +1,4 @@
-import { AccountData } from '../../domain/entity/account.entity.data';
+import { AccountData } from '../../../../domain/entity/account.entity.data';
 
 export type WithdrawAccountRestServiceParamsDTO = {
   nickname: string;
